+++ packages/select/src/components/ui/AsyncMultiSelect.tsx
@@ -1,26 +1,27 @@
 import * as React from "react";
 import { useMemo } from "react";
-import SelectComponent, { mergeStyles, Props } from "react-select";
+import { mergeStyles } from "react-select";
+import AsyncComponent, { Props } from "react-select/async";
 import { defaultSelectTheme, selectThemeDark } from "../../SelectTheme";
 import { createStylesFromTheme } from "../../util/StylesBuilder";
 
-export interface SelectProps<T = { label: string; value: string }>
-  extends Omit<Props<T, false>, "isMulti"> {
+export interface AsyncMultiSelectProps<T = { label: string; value: string }>
+  extends Omit<Props<T, true>, "isMulti"> {
   variant?: "dark" | "light";
 }
 
-export const Select = <T extends {}>({
+export const AsyncMultiSelect = <T extends {}>({
   variant = "light",
   styles,
   ...selectProps
-}: SelectProps<T>) => {
+}: AsyncMultiSelectProps<T>) => {
   const selectStyles = useMemo(() => {
-    const sourceStyles = createStylesFromTheme<T, false>(
+    const sourceStyles = createStylesFromTheme<T, true>(
       variant === "light" ? defaultSelectTheme : selectThemeDark
     );
 
     return styles ? mergeStyles(sourceStyles, styles) : sourceStyles;
   }, [variant, styles]);
 
-  return <SelectComponent styles={selectStyles} {...selectProps} />;
+  return <AsyncComponent styles={selectStyles} {...selectProps} />;
 };