+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 ## Select
 
 - Update `react-select` to 4.3.0.
-- New component `MultiSelect`.
+- New components `MultiSelect` and `AsyncMultiSelect` replaces `isMulti` prop.
+- New `ChipMultiSelect` component.
 
 ## Breaking changes